fix streak bias to give correct score
we need to be operating with only positive numbers since the datatype is
unsigned, but also we want an exact match to have 0 distance. fixed by
adding a temporary streak bias (the worst case scenario of streaks).
also we use @min to not add _too high_ a streak bias as this would
unnecessarily limit our useful operating range.